+++ src/runtime/tests/test_date.rs
@@ -234,7 +234,7 @@ impl<'x> Context<'x> {
             let bytes = self
                 .message
                 .raw_message
-                .get(header.offset_start..header.offset_end)?;
+                .get(header.offset_start as usize..header.offset_end as usize)?;
             if let HeaderValue::DateTime(dt) = MessageStream::new(bytes).parse_date() {
                 if dt.is_valid() {
                     return Some(Cow::Owned(dt));
